# openapi.yaml
openapi: 3.1.0
info:
  title: SPY Options API
  description: API for fetching SPY OHLC, IV, Greeks, Market Tide, Spot GEX, and Bid/Ask Volume data.
  version: 1.2.0
servers:
  - url: https://spy-options-api.onrender.com
paths:
...
  /api/spy/ohlc/daily:
    get:
      summary: Get full SPY OHLC intraday data
      operationId: getSPYOHLCFullDay
      description: Returns all 5-minute candles for the specified trading day.
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: The trading date to query (YYYY-MM-DD).
      responses:
        "200":
          description: Successfully retrieved full-day OHLC data.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    open:
                      type: number
                    high:
                      type: number
                    low:
                      type: number
                    close:
                      type: number
                    volume:
                      type: integer
                    total_volume:
                      type: integer
                    start_time:
                      type: string
                      format: date-time
                    end_time:
                      type: string
                      format: date-time
  /api/spy/intraday-summary:
    get:
      summary: Get SPY Intraday Summary
      description: Fetch SPY daily intraday summary including OHLC, VWAP, price structure, dealer exposure, market tide, and delta trends.
      parameters:
        - name: date
          in: query
          required: false
          description: Optional date to fetch intraday summary for (format YYYY-MM-DD). Defaults to today if not provided.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Successfully retrieved SPY intraday summary.
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  rolling_avg_18:
                    type: object
                    properties:
                      avg_close:
                        type: number
                      latest_close:
                        type: number
                  ohlc_summary:
                    type: object
                    properties:
                      open:
                        type: number
                      high:
                        type: number
                      low:
                        type: number
                      close:
                        type: number
                      total_volume:
                        type: integer
                      percent_from_open:
                        type: number
                  dealer_exposure:
                    type: object
                    properties:
                      spot_price:
                        type: number
                      gamma_oi:
                        type: number
                      charm_oi:
                        type: number
                      vanna_oi:
                        type: number
                      implied_volatility:
                        type: number
                  vwap:
                    type: number
                  price_structure:
                    type: object
                    properties:
                      retested_levels:
                        type: array
                        items:
                          type: number
                      support_zones:
                        type: array
                        items:
                          type: number
                      resistance_zones:
                        type: array
                        items:
                          type: number
                      consolidation_zones:
                        type: array
                        items:
                          type: object
                          properties:
                            range:
                              type: array
                              items:
                                type: number
                            duration:
                              type: string
                  market_tide:
                    type: object
                    properties:
                      last_3:
                        type: array
                        items:
                          type: object
                          properties:
                            timestamp:
                              type: string
                              format: date-time
                            net_call_premium:
                              type: number
                            net_put_premium:
                              type: number
                            net_volume:
                              type: integer
                      rolling_avg:
                        type: object
                        properties:
                          avg_net_call_premium_12:
                            type: number
                          avg_net_put_premium_12:
                            type: number
                          avg_net_volume_12:
                            type: number
                          avg_net_call_premium_48:
                            type: number
                          avg_net_put_premium_48:
                            type: number
                          avg_net_volume_48:
                            type: number
                          recorded_at:
                            type: string
                            format: date-time
                  market_tide_deltas:
                    type: array
                    items:
                      type: object
                      properties:
                        timestamp:
                          type: string
                          format: date-time
                        delta_call:
                          type: number
                        delta_put:
                          type: number
                        delta_volume:
                          type: integer
                        sentiment:
                          type: string
...
  /api/greek-exposure:
    get:
      summary: Fetch the last 5 Greek Exposure records for SPY or SPX
      operationId: getGreekExposure
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
            enum: [SPY, SPX]
          description: The symbol for which to fetch Greek Exposure data (SPY or SPX)
      responses:
        "200":
          description: Successful response with Greek Exposure data.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    symbol:
                      type: string
                    date:
                      type: string
                      format: date
                    call_charm:
                      type: number
                    call_delta:
                      type: number
                    call_gamma:
                      type: number
                    call_vanna:
                      type: number
                    put_charm:
                      type: number
                    put_delta:
                      type: number
                    put_gamma:
                      type: number
                    put_vanna:
                      type: number
                    recorded_at:
                      type: string
                      format: date-time
...
  /api/bid-shift-signals:
    get:
      summary: Filter bid shift signals by symbol and confidence
      operationId: getFilteredBidShiftSignals
      parameters:
        - name: symbol
          in: query
          required: false
          schema:
            type: string
          description: Optional symbol filter (e.g., SPY, SPX, QQQ)
        - name: confidence
          in: query
          required: false
          schema:
            type: string
            enum: [Low, Moderate, High]
          description: Optional confidence level filter
      responses:
        "200":
          description: Filtered list of bid shift signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BidShiftSignal'
  /api/darkpool/top:
    get:
      summary: Get top SPY dark pool levels
      description: |
        Returns the top aggregated SPY dark pool price levels for a given trading day.
        If no date is provided, the latest available trading day will be used as fallback.
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Specific trading date (YYYY-MM-DD) to fetch levels for.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Number of top levels to return.
      responses:
        '200':
          description: A list of top SPY dark pool price levels.
          content:
            application/json:
              schema:
                type: object
                properties:
                  trading_day:
                    type: string
                    format: date
                  fallback_date:
                    type: string
                    format: date
                  top_levels:
                    type: array
                    items:
                      type: object
                      properties:
                        price:
                          type: number
                        total_premium:
                          type: number
                        total_volume:
                          type: integer
                        total_size:
                          type: integer
                        trade_count:
                          type: integer
        '500':
          description: Internal Server Error
components:
  schemas:
    BidShiftSignal:
      type: object
      properties:
        symbol:
          type: string
        recorded_at:
          type: string
          format: date-time
        dominant_side:
          type: string
          enum: [CALL, PUT]
        previous_dominant_side:
          type: string
        shift_type:
          type: string
          enum: [PUT_TO_CALL, CALL_TO_PUT, NONE]
        continuation:
          type: boolean
        delta_confirmation:
          type: boolean
        confidence:
          type: string
          enum: [Low, Moderate, High]
        created_at:
          type: string
          format: date-time
...